+++ noodles-sam/src/lib.rs
@@ -41,4 +41,9 @@ pub mod record;
 pub use self::{header::Header, record::Record};
 
 #[cfg(feature = "async")]
-pub use self::r#async::io::{Reader as AsyncReader, Writer as AsyncWriter};
+#[deprecated(since = "0.65.0", note = "Use `sam::r#async::io::Reader` instead.")]
+pub use self::r#async::io::Reader as AsyncReader;
+
+#[cfg(feature = "async")]
+#[deprecated(since = "0.65.0", note = "Use `sam::r#async::io::Writer` instead.")]
+pub use self::r#async::io::Writer as AsyncWriter;